backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
authlib==1.3.0
httpx==0.27.0
pydantic-settings==2.3.4
